+++ app/controllers/concerns/backend/content_type_controller.rb
@@ -6,7 +6,6 @@ module Concerns
       included do
         layout 'backend/lightbox'
         before_action :find_model
-        helper_method :content_path
       end
 
       module ClassMethods
@@ -25,17 +24,9 @@ module Concerns
         @model = model.find(params[:id]).decorate
       end
 
-      def content_path
-        column = @model.column
-        path = "edit_translation_backend_#{column.row.rowable.class.to_s.downcase}_path"
-        # TODO this needs to be the locale that you were editing, not the interface locale.
-        # You can reproduce this by editing flexible content in NL when your interface is english.
-        send(path, column.row.rowable, @model.column.row.locale, anchor: "content-row-#{column.row.id}")
-      end
-
       def update
         if @model.update_attributes allowed_params
-          redirect_to content_path
+          render 'backend/lightbox_saved'
         else
           render :edit
         end